pyproject.toml

[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[project]
name = "SpeAcouPy"
version = "0.3.0"
description = "Loudspeaker Electro-Mechanical-Acoustical Modelling in Python"
authors = [{name="You"}]
requires-python = ">=3.9"
dependencies = ["numpy", "matplotlib", "pyyaml"]
readme = "README.md"
license = {text = "MIT"}
keywords = ["loudspeaker", "acoustics", "horn", "electroacoustics", "modeling", "impedance", "SPL"]

[tool.setuptools.packages.find]
where = [""]
include = ["speacoupy*"]

[project.scripts]
speacoupy = "speacoupy.cli:main"
